requirements.txt
# =================================================================
# Malaya LLM - Python Dependencies
# =================================================================

# --- Core Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# --- LLM & LangChain ---
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langsmith>=0.0.65
dspy-ai>=2.4.0  # Prompt optimization (additive to LangChain)

# --- gRPC Compatibility (DSPy/LiteLLM) ---
grpcio>=1.62.3,<1.68.0
grpcio-status>=1.62.3,<1.68.0

# --- RAG & Search ---
chromadb>=0.4.0
sentence-transformers>=2.2.0
rank_bm25>=0.2.2
tavily-python>=0.3.0

# --- HTTP Utilities ---
requests>=2.31.0

# --- Malaysian NLP ---
malaya>=5.0.0
faiss-cpu>=1.7.0  # Vector similarity search (FAISS)

# --- Security & Validation ---
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
jsonschema>=4.21.0
sentry-sdk>=1.45.0

# --- Data Processing ---
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers
# fasttext (Using wheel if needed, but adding here for tracking)
fasttext-wheel>=0.9.2

# --- Testing & Evaluation ---
pytest>=7.0.0
ragas>=0.1.0
httpx>=0.25.0  # For FastAPI testing
datasets>=2.14.0  # For Ragas evaluation

# --- Observability ---
slowapi>=0.1.9  # Rate limiting
prometheus-client>=0.20.0  # Metrics

# --- Voice (TTS/STT) ---
edge-tts>=6.1.0  # Malaysian TTS (free, Microsoft)
# openai-whisper>=20231117  # Optional: Local STT (uncomment for offline use)
malaya-speech>=1.3.0  # Malaysian ASR/TTS with dialect support

# --- Vision (Phase 2) ---
# qwen-vl-utils>=0.0.1  # For Qwen-VL vision model (uncomment when using)

# --- MCP (Model Context Protocol) ---
mcp>=0.1.0
